Restructure compaction picker code and rename some functions.
Summary:
Move universal compaction picker to separate files compaction_picker_universal.cc and compaction_picker_universal.h.
Rename some functions to make the code easier to understand.

Test Plan: Run all the test.
